requirements.txt
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.4
cachetools==3.1.1
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docopt==0.6.2
easydict==1.9
entrypoints==0.3
enum34==1.1.6
gitdb2==2.0.6
GitPython==3.0.5
google==2.0.2
google-auth==1.4.2
h5py==2.10.0
importlib-metadata==0.23
ipykernel==5.1.3
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonpickle==0.9.6
jsonschema==3.1.1
kiwisolver==1.1.0
lap==0.4.0
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
motmetrics==1.1.3
munch==2.5.0
munkres==1.1.2
nbconvert==5.6.1
nbformat==4.4.0
numpy==1.17.4
opencv-python==4.0.1.24
packaging==19.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==1.0.18
ptyprocess==0.6.0
py-cpuinfo==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.4.2
pyparsing==2.4.5
pyrsistent==0.15.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.5.5
rsa==4.0
sacred==0.8.0
scipy==1.3.2
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
smmap2==2.0.5
soupsieve==1.9.5
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
tornado==4.5.3
tqdm==4.38.0
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
